#90
separate methods from rolladie main class to `NewGame` class
-edit code in rolladie main to integrate NewGame class
